Add files via upload
In this file, My group developed a data ingestion pipeline to pull data from the Ontario Data Catalog, which we aimed at pulling data that contains information on violent crime rates in Ontario from 2008 to 2012. The dataset is available in CSV format named: "violent_crime_rates.csv"

In this file, our group used API technique to request connection between our terminal and the website source end. We used request library for that part and then standardized the content from the website in a json file. And we did a data extraction using the according attributes. Finally we did a data cleaning to prepare the data for the file to be ready to submit.