Add files via upload
This data contains information on violent crime rates in Ontario from 2008 to 2012. The dataset is available in CSV format.

The dataset includes data related to all types of violent crime under the Criminal Code. Specifically, it covers the following types of violent crime:

Homicide; Attempted murder; Sexual assault (levels 1-3); Assault; Robbery; Criminal Harassment; Making threats; Other Violent Assault

The columns of the dataset include information on the year of observation, the area in which the crime occurred, and the corresponding crime rates for different types of violent crime.